requirements.txt
openai-agents
python-docx
lxml
Pillow
pytest
importlib-metadata
python-dotenv
